Minor 2.5.1-beta
Entirely removed box_smoother.py and its functionality from the main detection loop. This is because, after further testing and evaluating other options for box stabilization, not only was the current method computationally heavy, but it only worked if there was only 1 object of a given class on screen. So, back to the drawing board a little bit there. Looking into SORT algorithms through Hungarian or Kalman filters.