2.6-beta
Major update that finally includes the basics behind how inferences are going to be measured within the pipeline.

I wrote out an entire document going over the equations related to how this system works and how it was implemented into this pipeline. Most of the changes are within the actual helper defining the new calls made into the loop. 

Initiates the aggregator class in the beginning of the detection loop (run_detection):
aggregator.push_frame_data(frame_t, males, females, objects)

Calls on the aggregator to initiate measuring over defined intervals:
aggregator = Aggregator(interval_sec=5, session_sec=10)

Includes a call to properly define the other classes as simply O, or objects:
other_counts = {
                'Feeder': sum(1 for b in current_boxes_list if names.get(b[5]) == 'Feeder'),
                'Main_Perch': sum(1 for b in current_boxes_list if names.get(b[5]) == 'Main_Perch'),
                'Wooden_Perch': sum(1 for b in current_boxes_list if names.get(b[5]) == 'Wooden_Perch'),
                'Sky_Perch': sum(1 for b in current_boxes_list if names.get(b[5]) == 'Sky_Perch'),
                'Nesting_Box': sum(1 for b in current_boxes_list if names.get(b[5]) == 'Nesting_Box')
            }

Finally, calls on the .csv generator and places the files into the appropriate folder within logs.
# determine scores folder
        scores_root = BASE_DIR / "logs" / "scores"  # ~/YOLO/logs/scores
        safe_src_name = re.sub(r"[^\w\-\.]", "_", raw_source_name)
        timestamp_str = datetime.now().strftime("%m-%d-%Y_%H-%M-%S")
        aggregator_folder = scores_root / safe_src_name / timestamp_str
        aggregator_folder.mkdir(parents=True, exist_ok=True)

        interval_file, session_file = aggregator.save_results(aggregator_folder)
        dashboard.log(f"[INFO] Interval results saved to {interval_file}")
        dashboard.log(f"[INFO] Session summary saved to {session_file}")

Here is the formal outline of the equations that this uses to work:

Defining this as a single function (I), defined by class (c).
Ic(ti,tj) → represents # of inferences (I) of class (c) in times (ti) & (tj)

If we wanted to take the overall count of a class (c) within the interval aggregate:
Cc,k = Ic(tk-1,tk)
so, we're taking a count (C) of class (c) within time interval (k) that is equal to the function (I)

As an example, if males were detected 150 times within a single interval:
CM,1 = 150(0,1) = 150

To provide a comparable result between males and females, we can normalize the result by setting it as a rate (R):
Rc,k = Cc,k/Δt = # of detections per minute

We can also set this as a rate (R) over the total number of frames processed (f) within the interval (k):
Rc,k = Cc,k/fk = # of detections within the processed frames of the interval

Now, we can make an easily digestible ratio between males and females:
Pk = CM,k/CF,k = M:F

Putting this all together, we can find the session aggregate total (Rc) finding the mean of each interval’s rate:
Rc = (1/n)(n∑k=1 Rc,k) 
we are finding the average of the total number of discrete observation rates by taking the sum of the rate (R) of class (c) over total time interval (n) within interval (k).

We should also go ahead and find the standard deviation of the rate to ensure its accuracy:
σRc = √(1/n−1)∑(Rc,k − Rc)2